pyproject.toml
[tool.poetry]
name = "imogen"
version = "0.1.0"
description = ""
authors = ["apsz3 <apsz3@git>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lark = "^1.2.2"
pillow = "^10.4.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
imogen = "imogen.cli:main"